+++ src/main/ts/ephox/alloy/positioning/mode/SelectionAnchor.ts
@@ -1,6 +1,6 @@
 import { FieldSchema } from '@ephox/boulder';
-import { Fun, Option, Struct, Unicode } from '@ephox/katamari';
-import { Element, Insert, Node, Position, Remove, Selection, Traverse, WindowSelection } from '@ephox/sugar';
+import { Option, Struct, Unicode } from '@ephox/katamari';
+import { Element, Insert, Node, Position, Remove, Selection, Traverse, WindowSelection, Compare } from '@ephox/sugar';
 
 import { Window } from '@ephox/dom-globals';
 
@@ -8,15 +8,13 @@ import * as Boxes from '../../alien/Boxes';
 import * as CssPosition from '../../alien/CssPosition';
 import * as Descend from '../../alien/Descend';
 import * as Fields from '../../data/Fields';
-import * as Bubble from '../layout/Bubble';
-import * as Layout from '../layout/Layout';
 import * as Origins from '../layout/Origins';
-import { SelectionAnchor, nu as NuAnchor, Anchoring } from './Anchoring';
+import { SelectionAnchor, Anchoring } from './Anchoring';
 import * as ContainerOffsets from './ContainerOffsets';
 import { AlloyComponent } from '../../api/component/ComponentApi';
-import { PositioningConfig } from '../../behaviour/positioning/PositioningTypes';
 import { SugarRange } from '../../alien/TypeDefinitions';
 import * as AnchorLayouts from './AnchorLayouts';
+import ContentAnchorCommon from './ContentAnchorCommon';
 
 const point: (element: Element, offset: number) => {element: () => Element; offset: () => number; } = Struct.immutable('element', 'offset');
 
@@ -55,67 +53,17 @@ const placement = (component: AlloyComponent, anchorInfo: SelectionAnchor, origi
         return rect;
       });
     }) as Option<{ left: () => number, top: () => number, width: () => number, height: () => number}>;
-    return optRect.map((rawRect) => {
-      // NOTE: We are going to have to do some interesting things to make inline toolbars not appear over the toolbar.
-      const point = CssPosition.screen(
-        Position(
-          Math.max(0, rawRect.left()),
-          Math.max(0, rawRect.top())
-        )
-      );
-
-      return Boxes.pointed(point, rawRect.width(), rawRect.height());
+    return optRect.bind((rawRect) => {
+      return ContentAnchorCommon.capRect(rawRect.left(), rawRect.top(), rawRect.width(), rawRect.height());
     });
   });
 
-  return selectionBox.map((box) => {
-    const points = [ rootPoint, box.point() ];
-    const topLeft = Origins.cata(origin,
-      () => {
-        return CssPosition.sumAsAbsolute(points);
-      },
-      () => {
-        return CssPosition.sumAsAbsolute(points);
-      },
-      () => {
-        return CssPosition.sumAsFixed(points);
-      }
-    );
-
-    const anchorBox = Boxes.rect(
-      topLeft.left(),
-      topLeft.top(),
-      box.width(),
-      box.height()
-    );
-
-    const targetElement: Option<Element> = getAnchorSelection(win, anchorInfo).bind((sel) => {
-      return Node.isElement(sel.start()) ? Option.some(sel.start()) : Traverse.parent(sel.start());
-    });
-
-    const layoutsLtr = (): Layout.AnchorLayout[] => {
-      return anchorInfo.showAbove ?
-        [ Layout.northeast, Layout.northwest, Layout.southeast, Layout.southwest, Layout.north, Layout.south ] :
-        [ Layout.southeast, Layout.southwest, Layout.northeast, Layout.northwest, Layout.south, Layout.south ];
-    };
-
-    const layoutsRtl = (): Layout.AnchorLayout[] => {
-      return anchorInfo.showAbove ?
-        [ Layout.northwest, Layout.northeast, Layout.southwest, Layout.southeast, Layout.north, Layout.south ] :
-        [ Layout.southwest, Layout.southeast, Layout.northwest, Layout.northeast, Layout.south, Layout.north ];
-    };
-
-    const elem = targetElement.getOr(component.element());
-    const layouts = AnchorLayouts.get(elem, anchorInfo, layoutsLtr(), layoutsRtl());
-
-    return NuAnchor({
-      anchorBox,
-      bubble: anchorInfo.bubble.getOr(Bubble.fallback()),
-      overrides: anchorInfo.overrides,
-      layouts,
-      placer: Option.none()
-    });
+  const targetElement: Option<Element> = getAnchorSelection(win, anchorInfo).bind((sel) => {
+    return Node.isElement(sel.start()) ? Option.some(sel.start()) : Traverse.parent(sel.start());
   });
+  const elem = targetElement.getOr(component.element());
+
+  return ContentAnchorCommon.calcNewAnchor(selectionBox, rootPoint, anchorInfo, origin, elem);
 };
 
 export default [