+++ src/test/ts/module/ephox/alloy/test/GuiSetup.ts
@@ -1,13 +1,13 @@
-import { Assertions, Pipeline, Step } from '@ephox/agar';
+import { Assertions, Pipeline, Step, TestLogs } from '@ephox/agar';
 import { Merger } from '@ephox/katamari';
 import { DomEvent, Element, Html, Insert, Remove } from '@ephox/sugar';
 import TestStore from './TestStore';
 import * as Attachment from '../../../../../../main/ts/ephox/alloy/api/system/Attachment';
 import * as Gui from '../../../../../../main/ts/ephox/alloy/api/system/Gui';
 import { document, console } from '@ephox/dom-globals';
-import { AgarLogs } from '@ephox/agar/lib/main/ts/ephox/agar/pipe/Pipe';
+import { AlloyComponent } from '../../../../../../main/ts/ephox/alloy/api/component/ComponentApi';
 
-const setup = (createComponent, f, success, failure) => {
+const setup = (createComponent, f: (doc: Element, body: Element, gui: Gui.GuiSystem, component: AlloyComponent, store) => Array<Step<any, any>>, success, failure) => {
   const store = TestStore();
 
   const gui = Gui.create();
@@ -27,7 +27,7 @@ const setup = (createComponent, f, success, failure) => {
     // tslint:disable-next-line
     // console.error(e);
     failure(e, logs);
-  }, AgarLogs.init());
+  }, TestLogs.init());
 };
 
 const mSetupKeyLogger = (body) => {